fixed auth permission denied error
