+++ extension/src/HomePopup/utils/lastVisited.ts
@@ -1,7 +1,7 @@
-import storage from "GlobalHelpers/chrome/storage";
 import { FIREBASE_DB_REF } from "@common/constants/firebase";
 import { STORAGE_KEYS } from "GlobalConstants";
-import { getFromFirebase } from "GlobalHelpers/firebase";
+import storage from "GlobalHelpers/chrome/storage";
+import { getFromFirebase } from "GlobalHelpers/firebase/database";
 import { LastVisited } from "../interfaces/lastVisited";
 
 export const syncLastVisitedToStorage = async () => {
