+++ extension/src/helpers/firebase/index.ts
@@ -1,0 +1,14 @@
+import { initializeApp } from "firebase/app";
+
+const firebaseApp = initializeApp({
+  apiKey: "AIzaSyDiMRlBhW36sLjEADoQj9T5L1H-hIDUAso",
+  authDomain: "bypass-links.firebaseapp.com",
+  databaseURL: "https://bypass-links.firebaseio.com/",
+  projectId: "bypass-links",
+  storageBucket: "bypass-links.appspot.com",
+  messagingSenderId: "603462573180",
+  appId: "1:603462573180:web:317c7f02f1f66b836f3df9",
+  measurementId: "G-ZGKPZFJ01Z",
+});
+
+export default firebaseApp;
